reworked relation to no longer being a property
reworked it due to introducing a lot of headache, instead relational connections will be handled through systems, as it should be.

also reworked a feature in the ECS that would require extreme amounts of memory to work. This will make it less performant, but it was not a realistic feature to have memory-wise.